Fix scorecard unification, standings ROUND error, playing hcp decimals, and week recap
- standings/individual: cast to ::numeric before ROUND to fix PostgreSQL error
- calc_playing_handicap: return int so playing hcp is always a whole number
- scores/view.html: remove inline hcp note, move stroke dots above score
- enter_week.html: move stroke dots above score, fix hcp decimal, fix cancel button
- sc-stroke-dots: always render span with min-height for consistent row height
- base.html: header logo links admin to admin panel
- week_summary: remove Match Results section, add View Scorecards button
- New route + template: /schedule/<season_id>/week/<week_num>/scorecards

+++ app/routes/standings.py
@@ -1035,9 +1035,9 @@ def individual(season_id):
             t.team_name,
             mr.role,
             COUNT(DISTINCT mr.matchup_id)        AS rounds_played,
-            ROUND(SUM(mr.total_points), 1)       AS total_points,
-            ROUND(SUM(mr.hole_points_won), 1)    AS hole_pts,
-            ROUND(SUM(mr.overall_point_won), 1)  AS overall_pts,
+            ROUND(SUM(mr.total_points)::numeric, 1)       AS total_points,
+            ROUND(SUM(mr.hole_points_won)::numeric, 1)    AS hole_pts,
+            ROUND(SUM(mr.overall_point_won)::numeric, 1)  AS overall_pts,
             SUM(CASE WHEN mr.overall_point_won >= 1.0 THEN 1 ELSE 0 END) AS wins,
             SUM(CASE WHEN mr.overall_point_won  = 0.5 THEN 1 ELSE 0 END) AS ties,
             SUM(CASE WHEN mr.overall_point_won  = 0.0 THEN 1 ELSE 0 END) AS losses
